requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
cryptography
bcrypt
python-jose[cryptography]
python-multipart==0.0.6
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv==1.0.0
requests==2.31.0
pyotp==2.9.0
qrcode==7.4.2
pillow